Create homeDrivePermissionCheck.ps1
Script is used to to test the permissions of home drives  for users. A list of permissions is obtained by using powerview, and then the script will attempt to write a file to each home drive pairing as the user currently logged in. If we are able to write to a drive that we should not, a windows explorer window is opened using invoke-item, screenshot is taken , and then the windows explorer window is closed. 

Logging goes to the $destinationFile variable for reporting purposes.